+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/base/ButtonAction.java
@@ -45,6 +45,8 @@ public class ButtonAction implements ASMSource {
     List<DisassemblyListener> listeners = new ArrayList<>();
 
     private String scriptName = "-";
+    
+    private String exportedScriptName = "-";
 
     private final DefineButtonTag buttonTag;
 
@@ -226,5 +228,15 @@ public class ButtonAction implements ASMSource {
         } catch (InterruptedException ex) {
             return new ArrayList<>();
         }
+    }
+    
+    @Override
+    public String getExportedScriptName() {
+        return exportedScriptName;
+    }
+    
+    @Override
+    public void setExportedScriptName(String scriptName) {
+        this.exportedScriptName = scriptName;
     }
 }